backend/requirements.txt
Django>=4.2

djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
psycopg2-binary>=2.9
django-environ>=0.11
Pillow>=10.0
django-ratelimit>=4.0
python-slugify>=8.0
openpyxl>=3.1
reportlab>=4.2
requests>=2.31
gunicorn>=21.2
 whitenoise
  
   